requirements.txt
streamlit
crewai
langchain
langchain-community
langchain-groq
faiss-cpu
python-dotenv
pypdf
transformers==4.41.1
sentence-transformers==2.7.0
crewai[litellm]